Moves notification data factory, renames renderers and data fields to be less alert oriented
