+++ src/sentry/notifications/platform/templates/__init__.py
@@ -1,10 +1,10 @@
 from .data_export import DataExportFailureTemplate, DataExportSuccessTemplate
-from .issue_alert import IssueAlertNotificationTemplate
+from .issue import IssueNotificationTemplate
 
 __all__ = (
     "DataExportSuccessTemplate",
     "DataExportFailureTemplate",
-    "IssueAlertNotificationTemplate",
+    "IssueNotificationTemplate",
 )
 # All templates should be imported here so they are registered in the notifications Django app.
 # See sentry/notifications/apps.py
